+++ server/server.go
@@ -20,6 +20,7 @@ import (
 	"github.com/influxdata/kapacitor/auth"
 	"github.com/influxdata/kapacitor/command"
 	iclient "github.com/influxdata/kapacitor/influxdb"
+	"github.com/influxdata/kapacitor/server/vars"
 	"github.com/influxdata/kapacitor/services/alert"
 	"github.com/influxdata/kapacitor/services/alerta"
 	"github.com/influxdata/kapacitor/services/azure"
@@ -62,7 +63,7 @@ import (
 	"github.com/influxdata/kapacitor/services/udp"
 	"github.com/influxdata/kapacitor/services/victorops"
 	"github.com/influxdata/kapacitor/uuid"
-	"github.com/influxdata/kapacitor/vars"
+	"github.com/influxdata/kapacitor/waiter"
 	"github.com/pkg/errors"
 )
 
@@ -86,6 +87,9 @@ type Server struct {
 	config *Config
 
 	err chan error
+
+	// clusterIDChanged signals when the cluster ID has changed.
+	clusterIDChanged *waiter.WaiterGroup
 
 	Commander command.Commander
 
@@ -139,19 +143,20 @@ func New(c *Config, buildInfo BuildInfo, logService logging.Interface) (*Server,
 	}
 	l := logService.NewLogger("[srv] ", log.LstdFlags)
 	s := &Server{
-		config:          c,
-		BuildInfo:       buildInfo,
-		dataDir:         c.DataDir,
-		hostname:        c.Hostname,
-		err:             make(chan error),
-		configUpdates:   make(chan config.ConfigUpdate, 100),
-		LogService:      logService,
-		MetaClient:      &kapacitor.NoopMetaClient{},
-		QueryExecutor:   &Queryexecutor{},
-		Logger:          l,
-		ServicesByName:  make(map[string]int),
-		DynamicServices: make(map[string]Updater),
-		Commander:       c.Commander,
+		config:           c,
+		BuildInfo:        buildInfo,
+		dataDir:          c.DataDir,
+		hostname:         c.Hostname,
+		err:              make(chan error),
+		configUpdates:    make(chan config.ConfigUpdate, 100),
+		LogService:       logService,
+		MetaClient:       &kapacitor.NoopMetaClient{},
+		QueryExecutor:    &Queryexecutor{},
+		Logger:           l,
+		ServicesByName:   make(map[string]int),
+		DynamicServices:  make(map[string]Updater),
+		Commander:        c.Commander,
+		clusterIDChanged: waiter.NewGroup(),
 	}
 	s.Logger.Println("I! Kapacitor hostname:", s.hostname)
 
@@ -171,7 +176,7 @@ func New(c *Config, buildInfo BuildInfo, logService logging.Interface) (*Server,
 
 	// Start Task Master
 	s.TaskMasterLookup = kapacitor.NewTaskMasterLookup()
-	s.TaskMaster = kapacitor.NewTaskMaster(kapacitor.MainTaskMaster, logService)
+	s.TaskMaster = kapacitor.NewTaskMaster(kapacitor.MainTaskMaster, vars.Info, logService)
 	s.TaskMaster.DefaultRetentionPolicy = c.DefaultRetentionPolicy
 	s.TaskMaster.Commander = s.Commander
 	s.TaskMasterLookup.Set(s.TaskMaster)
@@ -347,10 +352,16 @@ func (s *Server) appendInfluxDBService() error {
 	if err != nil {
 		return errors.Wrap(err, "failed to get http port")
 	}
-	srv, err := influxdb.NewService(c, httpPort, s.config.Hostname, s.config.HTTP.AuthEnabled, l)
+	srv, err := influxdb.NewService(c, httpPort, s.config.Hostname, vars.Info, s.config.HTTP.AuthEnabled, l)
 	if err != nil {
 		return err
 	}
+	w, err := s.newClusterIDChangedWaiter()
+	if err != nil {
+		return err
+	}
+	srv.ClusterIDWaiter = w
+
 	srv.HTTPDService = s.HTTPDService
 	srv.PointsWriter = s.TaskMaster
 	srv.LogService = s.LogService
@@ -673,7 +684,7 @@ func (s *Server) appendReportingService() {
 	c := s.config.Reporting
 	if c.Enabled {
 		l := s.LogService.NewLogger("[reporting] ", log.LstdFlags)
-		srv := reporting.NewService(c, l)
+		srv := reporting.NewService(c, vars.Info, l)
 
 		s.AppendService("reporting", srv)
 	}
@@ -852,6 +863,7 @@ func (s *Server) watchConfigUpdates() {
 // Close shuts down the meta and data stores and all services.
 func (s *Server) Close() error {
 	s.stopProfile()
+	s.clusterIDChanged.Stop()
 
 	// Close all services that write points first.
 	if err := s.HTTPDService.Close(); err != nil {
@@ -962,7 +974,12 @@ func (s *Server) SetClusterID(clusterID uuid.UUID) error {
 	}
 	s.ClusterID = clusterID
 	vars.ClusterIDVar.Set(s.ClusterID)
+	s.clusterIDChanged.Broadcast()
 	return nil
+}
+
+func (s *Server) newClusterIDChangedWaiter() (waiter.Waiter, error) {
+	return s.clusterIDChanged.NewWaiter()
 }
 
 // Service represents a service attached to the server.