+++ services/udp/service.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/influxdata/influxdb/models"
 	"github.com/influxdata/kapacitor/expvar"
-	"github.com/influxdata/kapacitor/vars"
+	"github.com/influxdata/kapacitor/server/vars"
 )
 
 const (