+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features
 
 - [#1413](https://github.com/influxdata/kapacitor/issues/1413): Add subscriptions modes to InfluxDB subscriptions.
+- [#1436](https://github.com/influxdata/kapacitor/issues/1436): Add linear fill support for QueryNode.
 
 ### Bugfixes